GithubLogin as a separate class
set

delete unnecessary awaits

abstract handlePopupLogin
