+++ e2e-tests/playwright/support/storage/githubAuth.ts
@@ -15,47 +15,16 @@ class GithubAuthStorage {
     return GithubAuthStorage.instance;
   }
 
-  /**
-   * Retrieves the StorageState associated with a username.
-   * @param userId - The username key.
-   * @returns The StorageState if found, otherwise null.
-   */
   public getStorageState(userId: string): StorageState | null {
     return this.storageStates.get(userId) || null;
   }
 
-  /**
-   * Stores a StorageState with the associated username.
-   * @param userId - The username key.
-   * @param state - The StorageState to store.
-   */
   public setStorageState(userId: string, state: StorageState): void {
     this.storageStates.set(userId, state);
   }
 
-  /**
-   * Removes the StorageState associated with a username.
-   * @param userId - The username key.
-   */
-  public removeStorageState(userId: string): void {
-    this.storageStates.delete(userId);
-  }
-
-  /**
-   * Checks if a StorageState exists for a given username.
-   * @param userId - The username key.
-   * @returns True if exists, otherwise false.
-   */
   public hasUsername(userId: string): boolean {
     return this.storageStates.has(userId);
-  }
-
-  /**
-   * Retrieves all stored usernames.
-   * @returns An array of usernames.
-   */
-  public getAllUsernames(): string[] {
-    return Array.from(this.storageStates.keys());
   }
 }
